major changes. reading and storing data works completely different but better now
alos added time import of raw data
improved reading of files

Because pickles are not long term stable (because they need a compatible environment as it was present upon creation), file structure was changed to json where all important info is stored. With this, we also have the meta information in a human readable form. which is a good thing. Import into python is also very easy.