improve saving and copying files.
copying: Now, tags are copied to a directory separate from images. Like this all data can be kept and the big images can only be on the computer for a limited amount of time

saving: when images are annotated, each annotation is saved to a database. Important: Lines are always appended. Thus no annotation which was done can ever be overwritten. Duplicates can be identified, because the tag id is stored. Unidentified duplicates however could occurr when the same analysis is done with different parameters