cre3ate multiIndex for Pandas, which is unfortunately not really helpful
I want timeseries objects in pandas because they have nice properties. But, they are not useful for the raw data of annotations, because for those i would need a multiindex, and this is limited and has no nice syntax. So the better option here is to create a grouped df for raw data and after processing create timeseries?!