image slicing version 0.1 is working satisfactory for now. In three example pictures the identification of the relevant sections of the image works reasonably well.
Core of the algorithm is the iteration of :

- mask construction based on minimum - mask - maximum filter application to create homogeneous regions of high contrast.

Ensurance of continous areas is done by detecting peaks or extrema in the area

and finally the new mask is created from this result and merged with the previous image.

Now we have a relatively homogeneous background to go for edge detection and merge this operation with movement. I can now try to detect the faintest movement and have this as a base set for "life" and match this with the edge detection.